move game stuff to the front end app
