--- static/index.html
@@ -1,18 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-  <head>
-    <meta charset="UTF-8" />
-    <title>Sample Site</title>
-    <script src="/static/map.js"></script>
-    <style>
-      body {
-        padding-top: 50px;
-      }
-    </style>
-  </head>
-  <body>
-    <div>
-      <canvas id="game"></canvas>
-    </div>
-  </body>
-</html>
